+++ packages/atomic/cypress/integration/query-summary-selectors.ts
@@ -1,3 +1,5 @@
+import {AriaLiveSelectors} from './aria-live-selectors';
+
 export const querySummaryComponent = 'atomic-query-summary';
 export const QuerySummarySelectors = {
   shadow: () => cy.get(querySummaryComponent).shadow(),
@@ -6,4 +8,5 @@ export const QuerySummarySelectors = {
     QuerySummarySelectors.shadow().find('[part="placeholder"]'),
   container: () => QuerySummarySelectors.shadow().find('[part="container"]'),
   duration: () => QuerySummarySelectors.shadow().find('[part="duration"]'),
+  ariaLive: () => AriaLiveSelectors.region('query-summary'),
 };