+++ packages/atomic/src/components/search/result-templates/atomic-result-template/atomic-result-template.tsx
@@ -8,6 +8,8 @@ import {makeMatchConditions} from '../../../common/result-template/result-templa
  * The `atomic-result-template` component determines the format of the query results, depending on the conditions that are defined for each template. A `template` element must be the child of an `atomic-result-template`, and either an `atomic-result-list` or `atomic-folded-result-list` must be the parent of each `atomic-result-template`.
  *
  * Note: Any `<script>` tags defined inside of a `<template>` element will not be executed when results are being rendered.
+ * @MapProp name: mustMatch;attr: must-match;docs: The field and values that define which result items the condition must be applied to. For example, a template with the following attribute only applies to result items whose `filetype` is `lithiummessage` or `YouTubePlaylist`: `must-match-filetype="lithiummessage,YouTubePlaylist"`;type: Record<string, string[]> ;default: {}
+ * @MapProp name: mustNotMatch;attr: must-not-match;docs: The field and values that define which result items the condition must not be applied to. For example, a template with the following attribute only applies to result items whose `filetype` is not `lithiummessage`: `must-not-match-filetype="lithiummessage";type: Record<string, string[]> ;default: {}
  */
 @Component({
   tag: 'atomic-result-template',
@@ -26,18 +28,8 @@ export class AtomicResultTemplate {
    */
   @Prop() public conditions: ResultTemplateCondition[] = [];
 
-  /**
-   * The field and values that define which result items the condition must be applied to.
-   *
-   * For example, a template with the following attribute only applies to result items whose `filetype` is `lithiummessage` or `YouTubePlaylist`: `must-match-filetype="lithiummessage,YouTubePlaylist"`
-   */
   @MapProp({splitValues: true}) public mustMatch: Record<string, string[]> = {};
 
-  /**
-   * The field and values that define which result items the condition must not be applied to.
-   *
-   * For example, a template with the following attribute only applies to result items whose `filetype` is not `lithiummessage`: `must-not-match-filetype="lithiummessage"`
-   */
   @MapProp({splitValues: true}) public mustNotMatch: Record<string, string[]> =
     {};